organize components directory
