+++ src/layouts/Layout.jsx
@@ -1,4 +1,4 @@
-import Navbar from "./Navbar";
+import Navbar from "../components/Navbar";
 
 export default function Layout({ children, session }) {
   return (
