+++ src/pages/_app.js
@@ -1,4 +1,4 @@
-import Layout from "@/components/Layout";
+import Layout from "@/layouts/Layout";
 import "@/styles/globals.css";
 import { SessionProvider } from "next-auth/react";
 import Head from "next/head";
@@ -14,14 +14,17 @@ export default function App({
     <>
       <Head>
         <title>Canine Assistants</title>
-        <meta name="description" content="Main dashboard page for Canine Assistants."/>
+        <meta
+          name="description"
+          content="Main dashboard page for Canine Assistants."
+        />
       </Head>
       <SessionProvider session={session}>
         {getLayout(
           <>
             <Toaster />
             <Component {...pageProps} />
-          </>
+          </>,
         )}
       </SessionProvider>
     </>
