+++ src/pages/dogs/[id]/forms/new.jsx
@@ -1,4 +1,4 @@
-import Form from "@/components/Form";
+import Form from "@/components/Form/Form";
 import { formActions } from "@/utils/formUtils";
 
 export default function NewForm() {
